==> src/de_analyses/17_compare_region_pairwise_results.Rmd <==
---
title: "compare single region to pairwise"
author: "Emma Jones"
date: "2023-09-26"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Compare region to pairwise specific gene lists

The purpose of this analysis is to determine how many DTU genes or DEGs are specific to a study design. Keep things things at the single vs others level is easier to interpret, while the pairwise finds more genes. However, the genes in common across pairwise comparisons are not the same as the single vs others genes.

### Load in Packages

```{r load in packages}
# load packages
suppressPackageStartupMessages({
  library(tidyverse)
  library(IsoformSwitchAnalyzeR)
  library(ComplexHeatmap)
  library(viridis)
  library(styler)
  library(lintr)
  library(here)
  library(eulerr)
  library(circlize)
})

# source functions
source("de_functions.R")
```

### Load in Data

Let's load in the data generated in script 15, that has all the dge, dte, and dtu lists.

```{r load in r environment data}
load(here("data", "comparison_gene_lists", "all_comparison_gene_lists.Rdata"))
```

## Examine if genes identified in 1 v all are in the pairwise gene lists

The purpose of this analysis is so see how much overlap there is between the pairwise and single region vs other comparisons for different brain regions.

#### Hippocampus

I'll start with the smallest gene list, hippocampus

```{r hippocampus dge}
# get length of single vs all
length(hippocampus_dge_genes) # 395

# first pairwise comparison
length(striatum_hippocampus_dge_genes) # 707
# how many genes are shared?
sum(hippocampus_dge_genes %in% striatum_hippocampus_dge_genes) # 237
# what percent is that total?
sum(hippocampus_dge_genes %in% striatum_hippocampus_dge_genes) /
  length(hippocampus_dge_genes) # 0.6

# second pairwise comparison
length(cerebellum_hippocampus_dge_genes) # 1935
# how many genes are shared?
sum(hippocampus_dge_genes %in% cerebellum_hippocampus_dge_genes) # 320
# what percent is that total?
sum(hippocampus_dge_genes %in% cerebellum_hippocampus_dge_genes) /
  length(hippocampus_dge_genes) # 0.81

# third pairwise comparison
length(cortex_hippocampus_dge_genes) # 349
# how many genes are shared?
sum(hippocampus_dge_genes %in% cortex_hippocampus_dge_genes) # 160
# what percent is that total?
sum(hippocampus_dge_genes %in% cortex_hippocampus_dge_genes) /
  length(hippocampus_dge_genes) # 0.41
```
Expected results:
[1] 395
[1] 707
[1] 237
[1] 0.6
[1] 1935
[1] 320
[1] 0.8101266
[1] 349
[1] 160
[1] 0.4050633

do for hippocampus dtu 

```{r hippocampus dtu}
# get length of single vs all
length(hippocampus_dtu_genes) # 44

# first pairwise comparison
length(hippocampus_striatum_dtu_genes) # 72
# how many genes are shared?
sum(hippocampus_dtu_genes %in% hippocampus_striatum_dtu_genes) # 1
# what percent is that total?
sum(hippocampus_dtu_genes %in% hippocampus_striatum_dtu_genes) /
  length(hippocampus_dtu_genes) # 0.02

# second pairwise comparison
length(cerebellum_hippocampus_dtu_genes) # 177
# how many genes are shared?
sum(hippocampus_dtu_genes %in% cerebellum_hippocampus_dtu_genes) # 3
# what percent is that total?
sum(hippocampus_dtu_genes %in% cerebellum_hippocampus_dtu_genes) /
  length(hippocampus_dtu_genes) # 0.068

# third pairwise comparison
length(cortex_hippocampus_dtu_genes) # 41
# how many genes are shared?
sum(hippocampus_dtu_genes %in% cortex_hippocampus_dtu_genes) # 2
# what percent is that total?
sum(hippocampus_dtu_genes %in% cortex_hippocampus_dtu_genes) /
  length(hippocampus_dtu_genes) # 0.045
```
Expected results:
[1] 44
[1] 72
[1] 1
[1] 0.02272727
[1] 177
[1] 3
[1] 0.06818182
[1] 41
[1] 2
[1] 0.04545455

#### Striatum

Then, we can try the same stats on the striatum

```{r striatum dge}
# get length of single vs all
length(striatum_dge_genes) # 744

# first pairwise comparison
length(striatum_hippocampus_dge_genes) # 707
# how many genes are shared?
sum(striatum_dge_genes %in% striatum_hippocampus_dge_genes) # 389
# what percent is that total?
sum(striatum_dge_genes %in% striatum_hippocampus_dge_genes) /
  length(striatum_dge_genes) # 0.52

# second pairwise comparison
length(cerebellum_striatum_dge_genes) # 2229
# how many genes are shared?
sum(striatum_dge_genes %in% cerebellum_striatum_dge_genes) # 657
# what percent is that total?
sum(striatum_dge_genes %in% cerebellum_striatum_dge_genes) /
  length(striatum_dge_genes) # 0.88

# third pairwise comparison
length(cortex_striatum_dge_genes) # 634
# how many genes are shared?
sum(striatum_dge_genes %in% cortex_striatum_dge_genes) # 351
# what percent is that total?
sum(striatum_dge_genes %in% cortex_striatum_dge_genes) /
  length(striatum_dge_genes) # 0.47
```
Expected results:
[1] 744
[1] 707
[1] 389
[1] 0.5228495
[1] 2229
[1] 657
[1] 0.8830645
[1] 634
[1] 351
[1] 0.4717742

And, do the same for striatum dtu 

```{r striatum dtu}
# get length of single vs all
length(striatum_dtu_genes) # 2

# first pairwise comparison
length(hippocampus_striatum_dtu_genes) # 72
# how many genes are shared?
sum(striatum_dtu_genes %in% hippocampus_striatum_dtu_genes) # 0
# what percent is that total?
sum(striatum_dtu_genes %in% hippocampus_striatum_dtu_genes) /
  length(striatum_dtu_genes) # 0

# second pairwise comparison
length(cerebellum_striatum_dtu_genes) # 558
# how many genes are shared?
sum(striatum_dtu_genes %in% cerebellum_striatum_dtu_genes) # 0
# what percent is that total?
sum(striatum_dtu_genes %in% cerebellum_striatum_dtu_genes) /
  length(striatum_dtu_genes) # 0

# third pairwise comparison
length(cortex_striatum_dtu_genes) # 276
# how many genes are shared?
sum(striatum_dtu_genes %in% cortex_striatum_dtu_genes) # 0
# what percent is that total?
sum(striatum_dtu_genes %in% cortex_striatum_dtu_genes) /
  length(striatum_dtu_genes) # 0
```
Expected results:
[1] 2
[1] 72
[1] 0
[1] 0
[1] 558
[1] 0
[1] 0
[1] 276
[1] 0
[1] 0

#### Cortex

Start with getting stats for cortex dge

```{r cortex dge}
# get length of single vs all
length(cortex_dge_genes) # 541

# first pairwise comparison
length(cortex_hippocampus_dge_genes) # 349
# how many genes are shared?
sum(cortex_dge_genes %in% cortex_hippocampus_dge_genes) # 157
# what percent is that total?
sum(cortex_dge_genes %in% cortex_hippocampus_dge_genes) /
  length(cortex_dge_genes) # 0.29

# second pairwise comparison
length(cerebellum_cortex_dge_genes) # 2201
# how many genes are shared?
sum(cortex_dge_genes %in% cerebellum_cortex_dge_genes) # 474
# what percent is that total?
sum(cortex_dge_genes %in% cerebellum_cortex_dge_genes) /
  length(cortex_dge_genes) # 0.87

# third pairwise comparison
length(cortex_striatum_dge_genes) # 634
# how many genes are shared?
sum(cortex_dge_genes %in% cortex_striatum_dge_genes) # 230
# what percent is that total?
sum(cortex_dge_genes %in% cortex_striatum_dge_genes) /
  length(cortex_dge_genes) # 0.42
```
Expected results:
[1] 541
[1] 349
[1] 157
[1] 0.2902033
[1] 2201
[1] 474
[1] 0.8761553
[1] 634
[1] 230
[1] 0.4251386

Next, get stats for cortex dtu 

```{r cortex dtu}
# get length of single vs all
length(cortex_dtu_genes) # 21

# first pairwise comparison
length(cortex_hippocampus_dtu_genes) # 41
# how many genes are shared?
sum(cortex_dtu_genes %in% cortex_hippocampus_dtu_genes) # 6
# what percent is that total?
sum(cortex_dtu_genes %in% cortex_hippocampus_dtu_genes) /
  length(cortex_dtu_genes) # 0.28

# second pairwise comparison
length(cerebellum_cortex_dtu_genes) # 483
# how many genes are shared?
sum(cortex_dtu_genes %in% cerebellum_cortex_dtu_genes) # 10
# what percent is that total?
sum(cortex_dtu_genes %in% cerebellum_cortex_dtu_genes) /
  length(cortex_dtu_genes) # 0.47

# third pairwise comparison
length(cortex_striatum_dtu_genes) # 276
# how many genes are shared?
sum(cortex_dtu_genes %in% cortex_striatum_dtu_genes) # 15
# what percent is that total?
sum(cortex_dtu_genes %in% cortex_striatum_dtu_genes) /
  length(cortex_dtu_genes) # 0.71
```
Expected results:
[1] 21
[1] 41
[1] 6
[1] 0.2857143
[1] 483
[1] 10
[1] 0.4761905
[1] 276
[1] 15
[1] 0.7142857

#### Cerebellum

And finally, get stats for the cerebellum.

First, we can get stats for cerebellum dge.

```{r cerebellum dge}
# get length of single vs all
length(cerebellum_dge_genes) # 2147

# first pairwise comparison
length(cerebellum_hippocampus_dge_genes) # 1935
# how many genes are shared?
sum(cerebellum_dge_genes %in% cerebellum_hippocampus_dge_genes) # 1518
# what percent is that total?
sum(cerebellum_dge_genes %in% cerebellum_hippocampus_dge_genes) /
  length(cerebellum_dge_genes) # 0.7

# second pairwise comparison
length(cerebellum_cortex_dge_genes) # 2201
# how many genes are shared?
sum(cerebellum_dge_genes %in% cerebellum_cortex_dge_genes) # 1730
# what percent is that total?
sum(cerebellum_dge_genes %in% cerebellum_cortex_dge_genes) /
  length(cerebellum_dge_genes) # 0.80

# third pairwise comparison
length(cerebellum_striatum_dge_genes) # 2229
# how many genes are shared?
sum(cerebellum_dge_genes %in% cerebellum_striatum_dge_genes) # 1666
# what percent is that total?
sum(cerebellum_dge_genes %in% cerebellum_striatum_dge_genes) /
  length(cerebellum_dge_genes) # 0.77
```
Expected results:
[1] 2147
[1] 1935
[1] 1518
[1] 0.7070331
[1] 2201
[1] 1730
[1] 0.8057755
[1] 2229
[1] 1666
[1] 0.7759665

Last, calculate stats for cerebellum dtu 

```{r cerebellum dtu}
# get length of single vs all
length(cerebellum_dtu_genes) # 129

# first pairwise comparison
length(cerebellum_hippocampus_dtu_genes) # 177
# how many genes are shared?
sum(cerebellum_dtu_genes %in% cerebellum_hippocampus_dtu_genes) # 42
# what percent is that total?
sum(cerebellum_dtu_genes %in% cerebellum_hippocampus_dtu_genes) /
  length(cerebellum_dtu_genes) # 0.32

# second pairwise comparison
length(cerebellum_cortex_dtu_genes) # 483
# how many genes are shared?
sum(cerebellum_dtu_genes %in% cerebellum_cortex_dtu_genes) # 104
# what percent is that total?
sum(cerebellum_dtu_genes %in% cerebellum_cortex_dtu_genes) /
  length(cerebellum_dtu_genes) # 0.80

# third pairwise comparison
length(cerebellum_striatum_dtu_genes) # 558
# how many genes are shared?
sum(cerebellum_dtu_genes %in% cerebellum_striatum_dtu_genes) # 106
# what percent is that total?
sum(cerebellum_dtu_genes %in% cerebellum_striatum_dtu_genes) /
  length(cerebellum_dtu_genes) # 0.82
```
Expected results:
[1] 129
[1] 177
[1] 42
[1] 0.3255814
[1] 483
[1] 104
[1] 0.8062016
[1] 558
[1] 106
[1] 0.8217054

## Calculate jaccard similarity

So how can we quantify these patterns? One way is to calculate the jaccard similarity.

```{r calculate jaccard similarity}
# list all single comparisons
aggregate_dtu_comparisons <- list(
  cerebellum_dtu_genes = cerebellum_dtu_genes,
  cortex_dtu_genes = cortex_dtu_genes,
  hippocampus_dtu_genes = hippocampus_dtu_genes,
  striatum_dtu_genes = striatum_dtu_genes
)

# list all pairwise comparisons
pairwise_dtu_comparisons <-
  list(
    cerebellum_cortex_dtu_genes = cerebellum_cortex_dtu_genes,
    cerebellum_hippocampus_dtu_genes = cerebellum_hippocampus_dtu_genes,
    cerebellum_striatum_dtu_genes = cerebellum_striatum_dtu_genes,
    cortex_striatum_dtu_genes = cortex_striatum_dtu_genes,
    cortex_hippocampus_dtu_genes = cortex_hippocampus_dtu_genes,
    hippocampus_striatum_dtu_genes = hippocampus_striatum_dtu_genes
  )

# list all so function can be applied
all_combinations <- expand_grid(names(
  aggregate_dtu_comparisons
), names(pairwise_dtu_comparisons))

# map function to all comparisions
all_combinations$jaccard_similarity <-
  map2(
    all_combinations$`names(aggregate_dtu_comparisons)`,
    all_combinations$`names(pairwise_dtu_comparisons)`,
    get_jaccard_similarity
  )

# make column as numeric for plotting
all_combinations$jaccard_similarity <-
  as.numeric(all_combinations$jaccard_similarity)
```

```{r make dtu heatmap}
better_rownames <- c(
  "Cerebellum vs. Striatum", "Cerebellum vs. Cortex",
  "Cerebellum vs. Hippocampus",
  "Striatum vs. Hippocampus", "Cortex vs. Striatum",
  "Cortex vs. Hippocampus"
)

col_fun <- colorRamp2(c(0, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7), magma(8))

for_heatmap_1 <- as.data.frame(pivot_wider(all_combinations,
  names_from =
    `names(aggregate_dtu_comparisons)`,
  values_from = jaccard_similarity
))

rownames(for_heatmap_1) <- for_heatmap_1$`names(pairwise_dtu_comparisons)`

for_heatmap_1 <- for_heatmap_1[, 2:5]

heatmap_1 <- Heatmap(as.matrix(for_heatmap_1),
  name = "Jaccard Similarity",
  col = col_fun,
  row_labels = better_rownames,
  row_order = c(3, 1, 2, 6, 4, 5),
  row_names_gp = gpar(fontface = "bold"),
  column_labels = c(
    "Cerebellum DTU", "Cortex DTU",
    "Hippocampus DTU", "Striatum DTU"
  ),
  column_names_gp = gpar(fontface = "bold")
)

heatmap_1
```

```{r make dge lists}
aggregate_dge_comparisons <- list(
  cerebellum_dge_genes = cerebellum_dge_genes,
  cortex_dge_genes = cortex_dge_genes,
  hippocampus_dge_genes = hippocampus_dge_genes,
  striatum_dge_genes = striatum_dge_genes
)

pairwise_dge_comparisons <- list(
  cerebellum_cortex_dge_genes = cerebellum_cortex_dge_genes,
  cerebellum_hippocampus_dge_genes = cerebellum_hippocampus_dge_genes,
  cerebellum_striatum_dge_genes = cerebellum_striatum_dge_genes,
  cortex_striatum_dge_genes = cortex_striatum_dge_genes,
  cortex_hippocampus_dge_genes = cortex_hippocampus_dge_genes,
  striatum_hippocampus_dge_genes = striatum_hippocampus_dge_genes
)

all_dge_combinations <- expand_grid(
  names(aggregate_dge_comparisons),
  names(pairwise_dge_comparisons)
)

all_dge_combinations$jaccard_similarity <-
  map2(
    all_dge_combinations$`names(aggregate_dge_comparisons)`,
    all_dge_combinations$`names(pairwise_dge_comparisons)`,
    get_jaccard_similarity
  )

all_dge_combinations$jaccard_similarity <-
  as.numeric(all_dge_combinations$jaccard_similarity)
```

```{r make dge heatmap}
col_fun <- colorRamp2(c(0, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7), magma(8))

for_heatmap_2 <- as.data.frame(pivot_wider(
  all_dge_combinations,
  names_from = `names(aggregate_dge_comparisons)`,
  values_from = jaccard_similarity
))

rownames(for_heatmap_2) <- for_heatmap_2$`names(pairwise_dge_comparisons)`

for_heatmap_2 <- for_heatmap_2[, 2:5]

heatmap_2 <- Heatmap(as.matrix(for_heatmap_2),
  name = "Jaccard Similarity",
  col = col_fun,
  row_labels = better_rownames,
  row_names_gp = gpar(fontface = "bold"),
  column_labels = c(
    "Cerebellum DGE", "Striatum DGE",
    "Cortex DGE", "Hippocampus DGE"
  ),
  column_names_gp = gpar(fontface = "bold")
)
heatmap_2
```


```{r show both heatmaps with same colors}
png(here("results", "finished_figures", "supp3.png"),
  width = 8, height = 6, units = "in", res = 300
)

draw(heatmap_1 + heatmap_2)

dev.off()

draw(heatmap_1 + heatmap_2)
```

#### Style and lint

```{r tidy script}
style_file("17_compare_region_pairwise_results.Rmd")
lint("17_compare_region_pairwise_results.Rmd",
  linters = linters_with_defaults(
    object_length_linter = NULL,
    object_name_linter = NULL,
    object_usage_linter = NULL
  )
)
```

#### Versions

```{r versions}
sessionInfo()
```
R version 4.3.1 (2023-06-16)
Platform: x86_64-pc-linux-gnu (64-bit)
Running under: Ubuntu 22.04.3 LTS

Matrix products: default
BLAS:   /usr/lib/x86_64-linux-gnu/openblas-pthread/libblas.so.3 
LAPACK: /usr/lib/x86_64-linux-gnu/openblas-pthread/libopenblasp-r0.3.20.so;  LAPACK version 3.10.0

locale:
 [1] LC_CTYPE=en_US.UTF-8       LC_NUMERIC=C               LC_TIME=en_US.UTF-8       
 [4] LC_COLLATE=en_US.UTF-8     LC_MONETARY=en_US.UTF-8    LC_MESSAGES=en_US.UTF-8   
 [7] LC_PAPER=en_US.UTF-8       LC_NAME=C                  LC_ADDRESS=C              
[10] LC_TELEPHONE=C             LC_MEASUREMENT=en_US.UTF-8 LC_IDENTIFICATION=C       

time zone: Etc/UTC
tzcode source: system (glibc)

attached base packages:
[1] grid      stats4    stats     graphics  grDevices utils     datasets  methods  
[9] base     

other attached packages:
 [1] circlize_0.4.15             eulerr_7.0.0                here_1.0.1                 
 [4] lintr_3.1.0                 styler_1.10.2               viridis_0.6.4              
 [7] viridisLite_0.4.2           ComplexHeatmap_2.16.0       IsoformSwitchAnalyzeR_2.1.3
[10] pfamAnalyzeR_1.0.1          sva_3.48.0                  genefilter_1.82.1          
[13] mgcv_1.9-0                  nlme_3.1-163                satuRn_1.8.0               
[16] DEXSeq_1.46.0               RColorBrewer_1.1-3          AnnotationDbi_1.62.2       
[19] DESeq2_1.40.2               SummarizedExperiment_1.30.2 GenomicRanges_1.52.0       
[22] GenomeInfoDb_1.36.3         IRanges_2.34.1              S4Vectors_0.38.1           
[25] MatrixGenerics_1.12.3       matrixStats_1.0.0           Biobase_2.60.0             
[28] BiocGenerics_0.46.0         BiocParallel_1.34.2         limma_3.56.2               
[31] lubridate_1.9.2             forcats_1.0.0               stringr_1.5.0              
[34] dplyr_1.1.3                 purrr_1.0.2                 readr_2.1.4                
[37] tidyr_1.3.0                 tibble_3.2.1                ggplot2_3.4.3              
[40] tidyverse_2.0.0            

loaded via a namespace (and not attached):
  [1] splines_4.3.1                 later_1.3.1                  
  [3] BiocIO_1.10.0                 bitops_1.0-7                 
  [5] filelock_1.0.2                R.oo_1.25.0                  
  [7] rex_1.2.1                     XML_3.99-0.14                
  [9] lifecycle_1.0.3               rprojroot_2.0.3              
 [11] edgeR_3.42.4                  doParallel_1.0.17            
 [13] processx_3.8.2                lattice_0.21-8               
 [15] ensembldb_2.24.0              backports_1.4.1              
 [17] magrittr_2.0.3                rmarkdown_2.24               
 [19] remotes_2.4.2.1               yaml_2.3.7                   
 [21] httpuv_1.6.11                 pbapply_1.7-2                
 [23] DBI_1.1.3                     abind_1.4-5                  
 [25] zlibbioc_1.46.0               R.cache_0.16.0               
 [27] R.utils_2.12.2                AnnotationFilter_1.24.0      
 [29] RCurl_1.98-1.12               rappdirs_0.3.3               
 [31] GenomeInfoDbData_1.2.10       annotate_1.78.0              
 [33] codetools_0.2-19              DelayedArray_0.26.7          
 [35] xml2_1.3.5                    tidyselect_1.2.0             
 [37] shape_1.4.6                   futile.logger_1.4.3          
 [39] locfdr_1.1-8                  BiocFileCache_2.8.0          
 [41] GenomicAlignments_1.36.0      jsonlite_1.8.7               
 [43] GetoptLong_1.0.5              ellipsis_0.3.2               
 [45] survival_3.5-7                iterators_1.0.14             
 [47] foreach_1.5.2                 tools_4.3.1                  
 [49] progress_1.2.2                Rcpp_1.0.11                  
 [51] glue_1.6.2                    gridExtra_2.3                
 [53] xfun_0.40                     withr_2.5.0                  
 [55] formatR_1.14                  BiocManager_1.30.22          
 [57] fastmap_1.1.1                 boot_1.3-28.1                
 [59] fansi_1.0.4                   callr_3.7.3                  
 [61] digest_0.6.33                 timechange_0.2.0             
 [63] R6_2.5.1                      mime_0.12                    
 [65] colorspace_2.1-0              biomaRt_2.56.1               
 [67] RSQLite_2.3.1                 R.methodsS3_1.8.2            
 [69] utf8_1.2.3                    generics_0.1.3               
 [71] data.table_1.14.8             tximeta_1.18.3               
 [73] rtracklayer_1.60.1            prettyunits_1.1.1            
 [75] httr_1.4.7                    S4Arrays_1.0.6               
 [77] pkgconfig_2.0.3               gtable_0.3.4                 
 [79] blob_1.2.4                    hwriter_1.3.2.1              
 [81] XVector_0.40.0                htmltools_0.5.6              
 [83] geneplotter_1.78.0            ProtGenerics_1.32.0          
 [85] clue_0.3-64                   scales_1.2.1                 
 [87] cyclocomp_1.1.1               png_0.1-8                    
 [89] knitr_1.44                    lambda.r_1.2.4               
 [91] rstudioapi_0.15.0             tzdb_0.4.0                   
 [93] reshape2_1.4.4                rjson_0.2.21                 
 [95] curl_5.0.2                    cachem_1.0.8                 
 [97] GlobalOptions_0.1.2           BiocVersion_3.17.1           
 [99] parallel_4.3.1                desc_1.4.2                   
[101] restfulr_0.0.15               pillar_1.9.0                 
[103] vctrs_0.6.3                   promises_1.2.1               
[105] dbplyr_2.3.3                  cluster_2.1.4                
[107] xtable_1.8-4                  tximport_1.28.0              
[109] evaluate_0.21                 VennDiagram_1.7.3            
[111] GenomicFeatures_1.52.2        cli_3.6.1                    
[113] locfit_1.5-9.8                compiler_4.3.1               
[115] futile.options_1.0.1          Rsamtools_2.16.0             
[117] rlang_1.1.1                   crayon_1.5.2                 
[119] ps_1.7.5                      plyr_1.8.8                   
[121] stringi_1.7.12                munsell_0.5.0                
[123] Biostrings_2.68.1             lazyeval_0.2.2               
[125] Matrix_1.6-1                  BSgenome_1.68.0              
[127] hms_1.1.3                     bit64_4.0.5                  
[129] KEGGREST_1.40.0               statmod_1.5.0                
[131] shiny_1.7.5                   interactiveDisplayBase_1.38.0
[133] AnnotationHub_3.8.0           memoise_2.0.1                
[135] bit_4.0.5                     xmlparsedata_1.0.5
