==> src/de_analyses/20_compare_enrichment_results.Rmd <==
---
title: "Compare Enrichement Results"
author: "Emma Jones"
date: "2023-12-13"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Compare Enrichment results across DGE, DTE, and DTU.

The purpose of this script is to get real statistics and values for directly comparing functional enrichment analysis results. It is dependent on scripts 00-18.
Run in github docker 1.7.

## Load in packages

```{r load in packages}
suppressPackageStartupMessages({
  library(tidyverse)
  library(gprofiler2)
  library(styler)
  library(lintr)
  library(here)
})
```

## Load in Data

```{r load in data}
cerebellum_cortex_dge_results <-
  read_csv(
    here("results/gprofiler2/dge/cerebellum_cortex_results.csv")
  )

cerebellum_cortex_dte_results <-
  read_csv(
    here("results/gprofiler2/dte/cerebellum_cortex_results.csv")
  )
    
cerebellum_cortex_dtu_results <-
  read_csv(
    here("results/gprofiler2/dtu/cerebellum_cortex_results.csv")
  )
```

## Get proportions of synaptic genes

Now, to do a very simple analysis, I will look at how many terms contain the phrase "synap"

```{r count rows and calculate proportions}
count_dge_rows <- sum(grepl("synap",
  cerebellum_cortex_dge_results$term_name,
  ignore.case = TRUE
))

count_dge_rows / nrow(cerebellum_cortex_dge_results) # 0.02870264

count_dte_rows <- sum(grepl("synap",
  cerebellum_cortex_dte_results$term_name,
  ignore.case = TRUE
))

count_dte_rows / nrow(cerebellum_cortex_dte_results) # 0.03167421

count_dtu_rows <- sum(grepl("synap",
  cerebellum_cortex_dtu_results$term_name,
  ignore.case = TRUE
))

count_dtu_rows / nrow(cerebellum_cortex_dtu_results) # 0.444444
```

#### Tidy script

```{r tidy script}
style_file("20_compare_enrichment_results.Rmd")
lint("20_compare_enrichment_results.Rmd",
  linters = linters_with_defaults(
    object_length_linter = NULL,
    object_name_linter = NULL,
    object_usage_linter = NULL
  )
)
```

#### Software versions

```{r versions}
sessionInfo()
```
R version 4.3.1 (2023-06-16)
Platform: x86_64-pc-linux-gnu (64-bit)
Running under: Ubuntu 22.04.3 LTS

Matrix products: default
BLAS:   /usr/lib/x86_64-linux-gnu/openblas-pthread/libblas.so.3 
LAPACK: /usr/lib/x86_64-linux-gnu/openblas-pthread/libopenblasp-r0.3.20.so;  LAPACK version 3.10.0

locale:
 [1] LC_CTYPE=en_US.UTF-8       LC_NUMERIC=C               LC_TIME=en_US.UTF-8       
 [4] LC_COLLATE=en_US.UTF-8     LC_MONETARY=en_US.UTF-8    LC_MESSAGES=en_US.UTF-8   
 [7] LC_PAPER=en_US.UTF-8       LC_NAME=C                  LC_ADDRESS=C              
[10] LC_TELEPHONE=C             LC_MEASUREMENT=en_US.UTF-8 LC_IDENTIFICATION=C       

time zone: Etc/UTC
tzcode source: system (glibc)

attached base packages:
[1] stats4    stats     graphics  grDevices utils     datasets  methods   base     

other attached packages:
 [1] gprofiler2_0.2.2            patchwork_1.1.3             DESeq2_1.40.2              
 [4] SummarizedExperiment_1.30.2 Biobase_2.60.0              MatrixGenerics_1.12.3      
 [7] matrixStats_1.0.0           GenomicRanges_1.52.0        GenomeInfoDb_1.36.3        
[10] IRanges_2.34.1              S4Vectors_0.38.1            BiocGenerics_0.46.0        
[13] cowplot_1.1.1               viridis_0.6.4               viridisLite_0.4.2          
[16] hexbin_1.28.3               here_1.0.1                  lintr_3.1.0                
[19] styler_1.10.2               data.table_1.14.8           lubridate_1.9.2            
[22] forcats_1.0.0               stringr_1.5.0               dplyr_1.1.3                
[25] purrr_1.0.2                 readr_2.1.4                 tidyr_1.3.0                
[28] tibble_3.2.1                ggplot2_3.4.3               tidyverse_2.0.0            

loaded via a namespace (and not attached):
 [1] bitops_1.0-7            gridExtra_2.3           remotes_2.4.2.1        
 [4] rlang_1.1.1             magrittr_2.0.3          compiler_4.3.1         
 [7] systemfonts_1.0.4       callr_3.7.3             vctrs_0.6.3            
[10] pkgconfig_2.0.3         crayon_1.5.2            fastmap_1.1.1          
[13] backports_1.4.1         XVector_0.40.0          labeling_0.4.3         
[16] utf8_1.2.3              rmarkdown_2.24          tzdb_0.4.0             
[19] ps_1.7.5                ragg_1.2.5              bit_4.0.5              
[22] xfun_0.40               zlibbioc_1.46.0         jsonlite_1.8.7         
[25] DelayedArray_0.26.7     BiocParallel_1.34.2     parallel_4.3.1         
[28] R6_2.5.1                stringi_1.7.12          Rcpp_1.0.11            
[31] knitr_1.44              R.utils_2.12.2          Matrix_1.6-1           
[34] R.cache_0.16.0          timechange_0.2.0        tidyselect_1.2.0       
[37] rstudioapi_0.15.0       abind_1.4-5             yaml_2.3.7             
[40] codetools_0.2-19        processx_3.8.2          lattice_0.21-8         
[43] withr_2.5.0             evaluate_0.21           desc_1.4.2             
[46] xml2_1.3.5              pillar_1.9.0            rex_1.2.1              
[49] plotly_4.10.2           generics_0.1.3          vroom_1.6.3            
[52] rprojroot_2.0.3         RCurl_1.98-1.12         hms_1.1.3              
[55] munsell_0.5.0           scales_1.2.1            glue_1.6.2             
[58] lazyeval_0.2.2          tools_4.3.1             locfit_1.5-9.8         
[61] grid_4.3.1              cyclocomp_1.1.1         colorspace_2.1-0       
[64] GenomeInfoDbData_1.2.10 cli_3.6.1               textshaping_0.3.6      
[67] fansi_1.0.4             S4Arrays_1.0.6          gtable_0.3.4           
[70] R.methodsS3_1.8.2       digest_0.6.33           xmlparsedata_1.0.5     
[73] htmlwidgets_1.6.2       farver_2.1.1            htmltools_0.5.6        
[76] R.oo_1.25.0             lifecycle_1.0.3         httr_1.4.7             
[79] bit64_4.0.5
